Add support for UDT data movers
Note this requires the user to download the UDT library
from sourceforge (http://udt.sourceforge.net), and install it somewhere.

Then, build the udtxfer app in src, after editing the Makfile in
that directory to point to the udt libraries.

Also, note that as UDT uses UDP, these UDP ports will also need to be
opened in firewalls.